# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EdgeEver API",
    "version": "0.1.0",
    "description": "REST API for EdgeEver notebooks, memos, resources, API tokens, and Agent integrations."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/memos": {
      "get": {
        "summary": "List or search memos",
        "parameters": [
          {
            "name": "notebookId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeDescendants",
            "in": "query",
            "description": "Include memos from all descendant notebooks when notebookId is set.",
            "schema": {
              "type": "string",
              "enum": ["1"]
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "trash",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["1"]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 100
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["updated-desc", "created-desc", "title-asc"],
              "default": "updated-desc"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["all", "tagged", "untagged", "pinned"],
              "default": "all"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Memo summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "memos": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MemoSummary"
                      }
                    },
                    "totalCount": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "nextCursor": {
                      "type": ["string", "null"]
                    }
                  },
                  "required": ["memos", "totalCount", "nextCursor"]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create memo",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoCreateInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/MemoResponse"
          }
        }
      }
    },
    "/api/v1/memos/batch/move": {
      "post": {
        "summary": "Move memos to notebook",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveMemosInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Move result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "moved": {
                      "type": "integer"
                    }
                  },
                  "required": ["ok", "moved"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memos/batch/delete": {
      "post": {
        "summary": "Delete memos",
        "description": "Soft-deletes active memos, or permanently deletes trashed memos when permanent is true.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteMemosInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Delete result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "deleted": {
                      "type": "integer"
                    }
                  },
                  "required": ["ok", "deleted"]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/memos/merge": {
      "post": {
        "summary": "Merge memos",
        "description": "Creates a new memo, joins source Markdown with separators, soft-deletes source memos, and re-points resources to the new memo.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeMemosInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/MemoResponse"
          }
        }
      }
    },
    "/api/v1/memos/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/MemoId"
        }
      ],
      "get": {
        "summary": "Get memo detail",
        "parameters": [
          {
            "name": "includeDeleted",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["1"]
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/MemoResponse"
          }
        }
      },
      "patch": {
        "summary": "Update memo",
        "description": "Use expectedRevision for optimistic concurrency. A mismatch returns revision_conflict.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoUpdateInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/MemoResponse"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "summary": "Soft-delete or permanently delete memo",
        "parameters": [
          {
            "name": "permanent",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["1"]
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          }
        }
      }
    },
    "/api/v1/memos/{id}/edit-sessions": {
      "parameters": [
        {
          "$ref": "#/components/parameters/MemoId"
        }
      ],
      "post": {
        "summary": "Start a bound memo edit session",
        "description": "Binds an edit session to the memo id, current revision, current content hash, and authenticated actor.",
        "responses": {
          "200": {
            "description": "Bound edit session",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "editSession": {
                      "$ref": "#/components/schemas/MemoEditSession"
                    }
                  },
                  "required": ["editSession"]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/memos/{id}/restore": {
      "parameters": [
        {
          "$ref": "#/components/parameters/MemoId"
        }
      ],
      "post": {
        "summary": "Restore memo from trash",
        "responses": {
          "200": {
            "$ref": "#/components/responses/MemoResponse"
          }
        }
      }
    },
    "/api/v1/memos/{id}/revisions": {
      "parameters": [
        {
          "$ref": "#/components/parameters/MemoId"
        }
      ],
      "get": {
        "summary": "List memo revisions",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Memo revisions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "revisions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MemoRevision"
                      }
                    }
                  },
                  "required": ["revisions"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/memos/{id}/revisions/{revisionId}/restore": {
      "parameters": [
        {
          "$ref": "#/components/parameters/MemoId"
        },
        {
          "name": "revisionId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "post": {
        "summary": "Restore a memo revision",
        "responses": {
          "200": {
            "$ref": "#/components/responses/MemoResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "edgeever_session"
      }
    },
    "parameters": {
...
      "MemoId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
      "Ok": {
        "description": "OK",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Ok"
            }
          }
        }
      },
      "Error": {
        "description": "API error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
      "MemoResponse": {
        "description": "Memo response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "memo": {
                  "$ref": "#/components/schemas/MemoDetail"
                }
              },
              "required": ["memo"]
            }
          }
        }
      },
...
    },
    "schemas": {
      "Ok": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": ["ok"]
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      },
...
      "MemoSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "notebookId": {
            "type": "string"
          },
          "title": {
            "type": ["string", "null"]
          },
          "excerpt": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isPinned": {
            "type": "boolean"
          },
          "isArchived": {
            "type": "boolean"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "revision": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": ["string", "null"],
            "format": "date-time"
          }
        },
        "required": ["id", "notebookId", "title", "excerpt", "tags", "isPinned", "isArchived", "isDeleted", "revision", "createdAt", "updatedAt", "deletedAt"]
      },
      "MemoDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/MemoSummary"
          },
          {
            "type": "object",
            "properties": {
              "contentJson": {
                "$ref": "#/components/schemas/TiptapDoc"
              },
              "contentMarkdown": {
                "type": "string"
              },
              "contentText": {
                "type": "string"
              },
              "contentHash": {
                "type": "string",
                "minLength": 64,
                "maxLength": 64
              },
              "sourceMemoIds": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "mergeSourceCount": {
                "type": "integer"
              },
              "mergedIntoMemoId": {
                "type": ["string", "null"]
              }
            },
            "required": ["contentJson", "contentMarkdown", "contentText", "contentHash", "sourceMemoIds", "mergeSourceCount", "mergedIntoMemoId"]
          }
        ]
      },
...
      "MemoCreateInput": {
        "type": "object",
        "properties": {
          "notebookId": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "maxLength": 160
          },
          "contentMarkdown": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": ["notebookId"]
      },
      "MemoUpdateInput": {
        "type": "object",
        "properties": {
          "expectedRevision": {
            "type": "integer",
            "minimum": 0
          },
          "expectedContentHash": {
            "type": "string",
            "minLength": 64,
            "maxLength": 64
          },
          "editSessionId": {
            "type": "string"
          },
          "notebookId": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "maxLength": 160
          },
          "isPinned": {
            "type": "boolean"
          },
          "contentJson": {
            "$ref": "#/components/schemas/TiptapDoc"
          },
          "contentMarkdown": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "MemoEditSession": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "memoId": { "type": "string" },
          "baseRevision": { "type": "integer", "minimum": 0 },
          "baseContentHash": { "type": "string", "minLength": 64, "maxLength": 64 },
          "expiresAt": { "type": "string", "format": "date-time" }
        },
        "required": ["id", "memoId", "baseRevision", "baseContentHash", "expiresAt"]
      },
      "MoveMemosInput": {
        "type": "object",
        "properties": {
          "memoIds": {
            "type": "array",
            "minItems": 1,
            "maxItems": 100,
            "items": {
              "type": "string"
            }
          },
          "notebookId": {
            "type": "string"
          }
        },
        "required": ["memoIds", "notebookId"]
      },
      "DeleteMemosInput": {
        "type": "object",
        "properties": {
          "memoIds": {
            "type": "array",
            "minItems": 1,
            "maxItems": 100,
            "items": {
              "type": "string"
            }
          },
          "permanent": {
            "type": "boolean"
          }
        },
        "required": ["memoIds"]
      },
      "MergeMemosInput": {
        "type": "object",
        "properties": {
          "memoIds": {
            "type": "array",
            "minItems": 2,
            "maxItems": 50,
            "items": {
              "type": "string"
            }
          },
          "notebookId": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "maxLength": 160
          }
        },
        "required": ["memoIds"]
      },
      "MemoRevision": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "memoId": {
            "type": "string"
          },
          "revision": {
            "type": "integer"
          },
          "title": {
            "type": ["string", "null"]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "contentMarkdown": {
            "type": "string"
          },
          "contentText": {
            "type": "string"
          },
          "contentHash": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "memoId", "revision", "title", "tags", "contentMarkdown", "contentText", "contentHash", "createdBy", "createdAt"]
      },
...
      "TiptapDoc": {
        "type": "object",
        "additionalProperties": true,
        "properties": {
          "type": {
            "type": "string",
            "const": "doc"
          },
          "content": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        },
        "required": ["type", "content"]
      },
...
    }
  }
}